Collector::AutoDBIC NOW STORES/VERIFIES DEPLOYED SCHEMA
When the AutoDBIC collector automatically generates and deploys the
audit schema it records a checksum/fingerprint which is then compared on
subsequent initializations for already populated audit schema dbs. This
protects schema mis-match if options are changed which results in a
different schema.

Also normalized 'get_dt' to obtain DateTime objects initialized in the
now configurable time_zone (defaults to 'local')