In prog Refactor: 'insert' tracking now in Storage Role
Tracking of 'insert' actions is now handled in the new Storage Role
(this is the easiest one). Still need to move update and delete, as well
as handling the cases where multiple rows are changed at once, such as
via direct ResultSet updates, and multiple inserts
(populate/insert_bulk). The multi-row support will be new (which was the
primary driver/purose of this refactoring)