In prog refactor to pre/post hook structure (changed 'insert')
In the progress of refactoring into a structure that can hook into
individual changes both before and after it is called. Converted over
'insert' hooks into new flow. This has no meaningful effect on the
front-end API yet, but it will allow for pre/post datapoint hooks to be
written into the structure to follow...

Note that the implementations in this commit are still preliminary and
will probably be simplified once the demo implementation is working