Reconsidered adding 'stage' to Datapoint - added more comments
Not going to make stage 'pre'/'post' a structural part of DataPoints
after all in favor of keeping DataPoints as simple organizational sugar.
Added more comments explaining this rationale. Pre/Post logic *still*
happens, but it is just restricted to Context objects, and if custom
Pre-change logic is needed it will just have to be implemented as custom
Context object(s).

Removed the unused 'stage' attr from the DataPoint class that had been
added in the previous commit