Get old 'init_sources' code working again for the role API
Very old code to be able initialize the state of existing rows has been
broken since a major refactoring (which was never fully completed) was
done to move code into a Storage role. This old code has been fixed to
be able to work again the way it did originally, with those tweaks
needed based on the subtle API changes. This is just being done to leave
this code in a more consistent state for if/when the rest of this task
is undertaken down the road. At least now the code paths work again w/o
throwing exceptions